- Resolved #113119 (screen result importer terminates prematurely 449). Chalked it up to excessive memory usage, by both JVM and postgres.
- ScreenResultImporter and ScreenResultParser now support incremental
importing of screen result files, by plate ranges.  This allows the importer
to handle a "manageable" size of data, without running out of memory, or
overloading postgres (which otherwise is being asked to complete the entire
import in a single transaction).

- Rewrote LibrariesDAOImpl.loadOrCreateLibraryWells().  I no longer understood
the comments we had in there, and wanted to have the code be written in a way
that made it very clear of its behavior w.r.t the Hibernate session; that is,
how the Library passed in is used when it's in either a transient, managed, or
persisted state.  The impl no longer relies upon catching of a Hibernate
exception to deal with a transient library object, avoids removed use of HQL.

- ResultValueType.resultValues map changed from Map<Well,ResultValue> to
Map<WellKey,ResultValue>.  This avoids having to have a full Well object
available to index into the map, which should optimize code that accesses this
map.

- WellKey is now an @Embeddable, so that it can be used as key type of
RVT.resultValues map.

- ResultValue.resultValueId now uses a 'seqhilo' ID generator, to optimize
bulk insert speed.  Resolves #113472 (ResultValue insert optimization).

- Updated unit tests.

git-svn-id: https://forge.abcd.harvard.edu/svn/screensaver/trunk@2105 80c535d0-b110-0410-b4de-9279f94142c3

+++ src/edu/harvard/med/screensaver/io/screenresults/ScreenResultImporter.java
@@ -27,6 +27,7 @@ import edu.harvard.med.screensaver.util.FileUtils;
 
 import org.apache.commons.cli.OptionBuilder;
 import org.apache.commons.cli.ParseException;
+import org.apache.commons.lang.math.IntRange;
 import org.apache.log4j.Logger;
 
 public class ScreenResultImporter
@@ -42,6 +43,9 @@ public class ScreenResultImporter
   public static final String[] SCREEN_OPTION = { "s", "screen" };
   public static final String[] IMPORT_OPTION = { "i", "import" };
   public static final String[] WELLS_OPTION = { "w", "wells" };
+  public static final String[] APPEND_OPTION = { "a", "append" };
+  public static final String[] PLATE_NUMBER_START_OPTION = { "sp", "start-plate" };
+  public static final String[] PLATE_NUMBER_END_OPTION = { "ep", "end-plate" };
 
   private static final String ERROR_ANNOTATED_WORKBOOK_FILE_EXTENSION = "errors.xls";
 
@@ -70,6 +74,19 @@ public class ScreenResultImporter
                                           .withDescription("the number of wells to print out")
                                           .withLongOpt(WELLS_OPTION[LONG_OPTION])
                                           .create(WELLS_OPTION[SHORT_OPTION]));
+    app.addCommandLineOption(OptionBuilder.withDescription("The first plate number to parse/import")
+                                           .hasArg()
+                                           .withArgName("#")
+                                           .withLongOpt(PLATE_NUMBER_START_OPTION[LONG_OPTION])
+                                           .create(PLATE_NUMBER_START_OPTION[SHORT_OPTION]));
+    app.addCommandLineOption(OptionBuilder.withDescription("The last plate number to parse/import")
+                                          .hasArg()
+                                          .withArgName("#")
+                                          .withLongOpt(PLATE_NUMBER_END_OPTION[LONG_OPTION])
+                                          .create(PLATE_NUMBER_END_OPTION[SHORT_OPTION]));
+    app.addCommandLineOption(OptionBuilder.withDescription("Append the specified range of plate numbers to an existing screen result.")
+                                          .withLongOpt(APPEND_OPTION[LONG_OPTION])
+                                          .create(APPEND_OPTION[SHORT_OPTION]));
     app.addCommandLineOption(OptionBuilder.withDescription("Import screen result into database if parsing is successful.  "
                                                            + "(By default, the parser only validates the input and then exits.)")
                                           .withLongOpt(IMPORT_OPTION[LONG_OPTION])
@@ -97,6 +114,18 @@ public class ScreenResultImporter
 
       final Integer wellsToPrint = app.getCommandLineOptionValue(WELLS_OPTION[SHORT_OPTION],
                                                                  Integer.class);
+
+      IntRange plateNumberRange = null;
+      if (app.isCommandLineFlagSet(PLATE_NUMBER_START_OPTION[SHORT_OPTION]) &&
+        app.isCommandLineFlagSet(PLATE_NUMBER_END_OPTION[SHORT_OPTION])) {
+        plateNumberRange = 
+          new IntRange(app.getCommandLineOptionValue(PLATE_NUMBER_START_OPTION[SHORT_OPTION], Integer.class),
+                       app.getCommandLineOptionValue(PLATE_NUMBER_END_OPTION[SHORT_OPTION], Integer.class));
+        log.info("will parse/load plates " + plateNumberRange);
+      }
+      final IntRange finalPlateNumberRange = plateNumberRange;
+      final boolean append = app.isCommandLineFlagSet(APPEND_OPTION[SHORT_OPTION]); 
+
       final Screen finalScreen = screen;
       final ScreenResultParser finalScreenResultParser = screenResultParser;
       dao.doInTransaction(new DAOTransaction() {
@@ -105,12 +134,19 @@ public class ScreenResultImporter
           dao.reattachEntity(finalScreen);
 
           if (finalScreen.getScreenResult() != null) {
-            log.info("deleting existing screen result for " + finalScreen);
-            screenResultsDao.deleteScreenResult(finalScreen.getScreenResult());
+            if (append) {
+              log.info("appending existing screen result (loading existing screen result data)");
+              dao.need(finalScreen.getScreenResult(), "resultValueTypes.resultValues");
+            }
+            else {
+              log.info("deleting existing screen result for " + finalScreen);
+              screenResultsDao.deleteScreenResult(finalScreen.getScreenResult());
+            }
           }
 
           ScreenResult screenResult = finalScreenResultParser.parse(finalScreen,
-                                                                    inputFile);
+                                                                    inputFile,
+                                                                    finalPlateNumberRange);
           if (finalScreenResultParser.getHasErrors()) {
             log.error("Errors encountered during parse:");
             for (WorkbookParseError error : finalScreenResultParser.getErrors()) {
@@ -153,6 +189,7 @@ public class ScreenResultImporter
       log.error("application error: " + e.getMessage());
       e.printStackTrace();
     }
+    log.info("Exiting.");
   }
 
   private static Screen findScreenOrExit(CommandLineApplication app) throws ParseException
